requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alpaca-trade-api==3.2.0
appdirs @ file:///home/conda/feedstock_root/build_artifacts/appdirs_1733753955715/work
attrs==25.3.0
beautifulsoup4 @ file:///home/conda/feedstock_root/build_artifacts/beautifulsoup4_1756094236201/work
Brotli @ file:///D:/bld/brotli-split_1756599174049/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1754231422783/work/certifi
cffi @ file:///D:/bld/cffi_1756808282238/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1754767332901/work
curl_cffi==0.13.0
deprecation==2.1.0
frozendict @ file:///D:/bld/frozendict_1756047926195/work
frozenlist==1.7.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
html5lib @ file:///home/conda/feedstock_root/build_artifacts/html5lib_1734075161666/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
lxml==6.0.1
msgpack==1.0.3
multidict==6.6.4
multitasking @ file:///home/conda/feedstock_root/build_artifacts/multitasking_1753115232125/work
numpy @ file:///D:/bld/numpy_1707225528937/work/dist/numpy-1.26.4-cp311-cp311-win_amd64.whl#sha256=537f0573fefb1e93b3633e8a79102e6b6d2911e09e0a90008993efef4242448e
packaging==25.0
pandas @ file:///D:/bld/pandas_1755778903675/work
pandas_ta @ file:///home/conda/feedstock_root/build_artifacts/pandas-ta_1737268314066/work
peewee @ file:///D:/bld/bld/rattler-build_peewee_1756672097/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1756227402/work
propcache==0.3.2
protobuf==6.32.1
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
PySocks @ file:///D:/bld/pysocks_1733217287171/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.1.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML==6.0.1
requests==2.32.3
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
soupsieve @ file:///home/conda/feedstock_root/build_artifacts/soupsieve_1756330469801/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
upstox-instrument-query==0.2.1
upstox-python-sdk==2.17.0
urllib3==1.26.20
uuid==1.30
webencodings @ file:///home/conda/feedstock_root/build_artifacts/webencodings_1733236011802/work
websocket-client==1.8.0
websockets==15.0.1
win_inet_pton @ file:///D:/bld/win_inet_pton_1733130564612/work
yarl==1.20.1
yfinance==0.2.66
zstandard==0.25.0
